refactor(e2e): modal frame tests
