+++ e2e/modal-navigation/e2e/e2e-spec-app-root/app-root-modal-frame-background.e2e-spec.ts
@@ -1,0 +1,60 @@
+import { AppiumDriver, createDriver } from "nativescript-dev-appium";
+import { Screen } from "../screen"
+import { modalFrameBackground,
+    testSecondPageBackground,
+    testNestedModalFrameBackground,
+    testNestedModalPageBackground
+} from "../e2e-spec-shared/modal-frame"
+
+describe("app root modal frame background scenarios", () => {
+
+    let driver: AppiumDriver;
+    let screen: Screen;
+
+    before(async () => {
+        driver = await createDriver();
+        screen = new Screen(driver);
+        // should load app root
+        await screen.loadedHome();
+    });
+
+    beforeEach(async function () {
+        await screen.loadModalFrame();
+    });
+
+    afterEach(async function () {
+        if (this.currentTest.state === "failed") {
+            await driver.logPageSource(this.currentTest.title);
+            await driver.logScreenshot(this.currentTest.title);
+            await driver.resetApp();
+        }
+    });
+
+    after(async () => {
+        // should close page with frame
+        await screen.closeModal();
+        await screen.loadedHome();
+        await driver.quit();
+        console.log("Quit driver!");
+    });
+
+    it("should run modal page with frame in background", async () => {
+        await modalFrameBackground(driver, screen);
+    });
+
+    it("should navigate to second page, run in background, go back", async () => {
+        await testSecondPageBackground(driver, screen);
+    });
+
+    it("should show nested modal page with frame, run in background, close", async () => {
+        await testNestedModalFrameBackground(driver, screen);
+    });
+
+    it("should show nested modal page, run in background, close", async () => {
+        await testNestedModalPageBackground(driver, screen);
+    });
+
+    it("should navigate to second page, run in background, go back", async () => {
+        await testSecondPageBackground(driver, screen);
+    });
+});
