+++ e2e/modal-navigation/e2e/e2e-spec-app-root/app-root-modal-frame.e2e-spec.ts
@@ -1,0 +1,52 @@
+import { AppiumDriver, createDriver } from "nativescript-dev-appium";
+import { Screen } from "../screen"
+import { testSecondPage, testNestedModalFrame, testNestedModalPage } from "../e2e-spec-shared/modal-frame"
+
+describe("app root modal frame scenarios", () => {
+
+    let driver: AppiumDriver;
+    let screen: Screen;
+
+    before(async () => {
+        driver = await createDriver();
+        screen = new Screen(driver);
+        // should load app root
+        await screen.loadedHome();
+    });
+
+    beforeEach(async function () {
+        await screen.loadModalFrame();
+    });
+
+    afterEach(async function () {
+        if (this.currentTest.state === "failed") {
+            await driver.logPageSource(this.currentTest.title);
+            await driver.logScreenshot(this.currentTest.title);
+            await driver.resetApp();
+        }
+    });
+
+    after(async () => {
+        // should close page with frame
+        await screen.closeModal();
+        await screen.loadedHome();
+        await driver.quit();
+        console.log("Quit driver!");
+    });
+
+    it("should navigate to second page, go back", async () => {
+        await testSecondPage(screen);
+    });
+
+    it("should show nested modal page with frame, close", async () => {
+        await testNestedModalFrame(screen);
+    });
+
+    it("should show nested modal page, close", async () => {
+        await testNestedModalPage(screen);
+    });
+
+    it("should navigate to second page, go back", async () => {
+        await testSecondPage(screen);
+    });
+});
